+++ src/index.css
@@ -6,6 +6,8 @@ body {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   background-color: rgb(236, 234, 234);
+  height: 100%;
+  width: 100%;
 }
 
 code {
@@ -109,4 +111,19 @@ code {
   margin-top: 35px;
   font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
   font-weight: 400;
+}
+.login-page {
+  align-items: center;
+  display: flex;
+  justify-content: center;
+  height: 100%;
+  width: 100%;
+}
+.login-form {
+  width: 330px;
+  position: absolute;
+  top: 200px;
+  padding: 35px 20px;
+  background-color: white;
+  border-top: turquoise solid 10px;
 }
